Some more renaming / repackaging
