+++ Core/src/main/java/com/pro_crafting/mc/worldfuscator/WorldFuscator.java
@@ -1,6 +1,6 @@
 package com.pro_crafting.mc.worldfuscator;
 
-import com.pro_crafting.mc.worldfuscator.example.BlockDisguiser;
+import com.pro_crafting.mc.worldfuscator.engine.BlockDisguiser;
 import com.comphenix.protocol.ProtocolLibrary;
 import com.comphenix.protocol.utility.MinecraftVersion;
 import org.bukkit.Bukkit;
